refactor: split up main/ryzenadj, add unit test
