+++ src/main/ryzenadj/index.ts
@@ -1,0 +1,5 @@
+// normally avoid index files. only breaking this module up for test mocking purposes,
+// so it makes sense to export all functions from one file.
+export * from './api'
+export * from './command'
+export * from './parser'
